+++ Softeq.XToolkit.WhiteLabel.Essentials.Droid/Softeq.XToolkit.WhiteLabel.Essentials.Droid.csproj
@@ -43,6 +43,7 @@
     </ItemGroup>
     <ItemGroup>
         <Compile Include="FullScreenImage\FullScreenImageDialogFragment.cs" />
+        <Compile Include="FullScreenImage\GestureListener.cs" />
         <Compile Include="ImagePicker\DroidImagePickerResult.cs" />
         <Compile Include="ImagePicker\DroidImagePickerService.cs" />
         <Compile Include="ImagePicker\ImagePicker.cs" />
